+++ CosmicBot/Commands/MinecraftTaskCommandModule.cs
@@ -1,4 +1,5 @@
-﻿using CosmicBot.Service;
+﻿using CosmicBot.DiscordResponse;
+using CosmicBot.Service;
 using Discord;
 using Discord.Interactions;
 
@@ -17,24 +18,48 @@ namespace CosmicBot.Commands
         [SlashCommand("add", "Add a scheduled command to run on your server. Start time: hh:mm:ss. Interval [d.]hh:mm:ss.")]
         public async Task AddTask(string serverId, string taskName, string commands, string startTime, string interval)
         {
+            if (!HasChannelPermissions())
+            {
+                await Respond(new MessageResponse("I don't have valid permissions in this channel", ephemeral: true));
+                return;
+            }
+
             await Respond(await _service.AddTask(serverId, taskName, commands, startTime, interval));
         }
 
         [SlashCommand("update", "Update a scheduled command on your server. Start time: hh:mm:ss. Interval [d.]hh:mm:ss.")]
         public async Task UpdateTask(string taskId, string? taskName, string? commands, string? startTime, string? interval)
         {
+            if (!HasChannelPermissions())
+            {
+                await Respond(new MessageResponse("I don't have valid permissions in this channel", ephemeral: true));
+                return;
+            }
+
             await Respond(await _service.UpdateTask(taskId, taskName, commands, startTime, interval));
         }
 
         [SlashCommand("remove", "Delete a scheduled command on your server.")]
         public async Task DeleteTask(string taskId)
         {
+            if (!HasChannelPermissions())
+            {
+                await Respond(new MessageResponse("I don't have valid permissions in this channel", ephemeral: true));
+                return;
+            }
+
             await Respond(await _service.RemoveTask(taskId));
         }
 
         [SlashCommand("list", "List your current scheduled tasks for a server")]
         public async Task ListTasks(string serverId)
         {
+            if (!HasChannelPermissions())
+            {
+                await Respond(new MessageResponse("I don't have valid permissions in this channel", ephemeral: true));
+                return;
+            }
+
             await Respond(await _service.ListTasks(serverId));
         }
     }
